file_update.sh
This script will update the template.file file with the appropriate value of `facter -p widget` from each remote server listed in server_update_list file. It keeps track of successes and failures and reports this back to you. The script will also return a list of failed IP addresses.